+++ src/resources/views/parcels.blade.php
@@ -4,8 +4,8 @@
     <div class="response-wrapper">
         <div class="response">
             <h2>Parcels</h2>
-            <button class="btn btn-primary" href="#" data-bs-toggle="modal" data-bs-target="#createParcel"
-                class="nav-link">Create
+            <button class="btn btn-primary mb-3" href="#" data-bs-toggle="modal" data-bs-target="#createParcel"
+                class="nav-link">Create a
                 Parcel</button>
         </div>
     </div>